refactor(tvmaze): move utility functions from tvmazeModel.ts to tvMazeUtils.ts (closes #64)
